Добавил файд init.php
