+++ sign-up.php
@@ -23,8 +23,8 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     $file_url = MOCK_IMG;
     // Валидация на загрузку файла с картинкой лота
     // Проверяем есть ли каталог для загрузки картинок на сервере
-    if(!file_exists('/upload/')){
-        mkdir('/upload/');
+    if(!file_exists(UPLOAD_LOCAL_DIR)){
+        mkdir(UPLOAD_LOCAL_DIR);
     }
     if (isset($_FILES['img-avatar']['name']) && !empty($_FILES['img-avatar']['name'])) {
 
@@ -39,7 +39,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
             $finfo = finfo_open(FILEINFO_MIME_TYPE);
             $file_type = finfo_file($finfo, $file_tmp_name);
             $file_name_uniq = uniqid('avatar-') . '.' . pathinfo($file_name , PATHINFO_EXTENSION);
-            $file_url = '/upload/' . trim($file_name_uniq);
+            $file_url = UPLOAD_LOCAL_DIR . trim($file_name_uniq);
             // Перемещение загруженного файла в папку сайта
             move_uploaded_file($file_tmp_name, UPLOAD_DIR . $file_name_uniq);
         }
